Updates for refactoriung spark connector
