+++ docs/sdk/queries/databricks/databricks-sql.md
@@ -144,8 +144,6 @@ This is a code example of the RTDIP SDK Interpolate function. You will need to r
 from rtdip_sdk.authentication import authenticate as auth
 from rtdip_sdk.connectors import DatabricksSQLConnection
 from rtdip_sdk.queries import interpolate
-from rtdip_sdk.odbc.db_sql_connector import DatabricksSQLConnection
-from rtdip_sdk.functions import interpolate
 
 authentication = auth.DefaultAuth().authenticate()
 access_token = authentication.get_token("2ff814a6-3304-4ab8-85cb-cd0e6f879c1d/.default").token
